+++ egui_commonmark_macros/tests/fail/incorrect_type.rs
@@ -4,5 +4,5 @@ use egui_commonmark_macros::commonmark;
 fn main() {
     let mut cache = egui_commonmark_backend::CommonMarkCache::default();
     let x = 3;
-    commonmark!("a", x, &mut cache, "# Hello");
+    commonmark!(x, &mut cache, "# Hello");
 }