+++ egui_commonmark_macros/tests/pass/book.rs
@@ -6,45 +6,44 @@ fn main() {
     let mut cache = egui_commonmark_backend::CommonMarkCache::default();
     __run_test_ui(|ui| {
         commonmark_str!(
-            "n2",
             ui,
             &mut cache,
             "../../../../egui_commonmark/examples/markdown/hello_world.md"
         );
 
         commonmark_str!(
-            "n3",
             ui,
             &mut cache,
             "../../../../egui_commonmark/examples/markdown/headers.md"
         );
 
         commonmark_str!(
-            "n6",
             ui,
             &mut cache,
             "../../../../egui_commonmark/examples/markdown/lists.md"
         );
 
         commonmark_str!(
-            "n7",
             ui,
             &mut cache,
             "../../../../egui_commonmark/examples/markdown/code-blocks.md"
         );
 
         commonmark_str!(
-            "n9",
             ui,
             &mut cache,
             "../../../../egui_commonmark/examples/markdown/blockquotes.md"
         );
 
         commonmark_str!(
-            "n11",
             ui,
             &mut cache,
             "../../../../egui_commonmark/examples/markdown/tables.md"
+        );
+        commonmark_str!(
+            ui,
+            &mut cache,
+            "../../../../egui_commonmark/examples/markdown/definition_list.md"
         );
     });
 }